+++ cmd/repo-updater/repos/gitlab.go
@@ -15,17 +15,18 @@ import (
 	"github.com/sourcegraph/sourcegraph/pkg/httpcli"
 	"github.com/sourcegraph/sourcegraph/pkg/jsonc"
 	"github.com/sourcegraph/sourcegraph/schema"
-	log15 "gopkg.in/inconshreveable/log15.v2"
+	"gopkg.in/inconshreveable/log15.v2"
 )
 
 // A GitLabSource yields repositories from a single GitLab connection configured
 // in Sourcegraph via the external services configuration.
 type GitLabSource struct {
-	svc     *ExternalService
-	config  *schema.GitLabConnection
-	exclude map[string]bool
-	baseURL *url.URL // URL with path /api/v4 (no trailing slash)
-	client  *gitlab.Client
+	svc                 *ExternalService
+	config              *schema.GitLabConnection
+	exclude             map[string]bool
+	baseURL             *url.URL // URL with path /api/v4 (no trailing slash)
+	nameTransformations reposource.NameTransformations
+	client              *gitlab.Client
 }
 
 // NewGitLabSource returns a new GitLabSource from the given external service.
@@ -73,12 +74,19 @@ func newGitLabSource(svc *ExternalService, c *schema.GitLabConnection, cf *httpc
 		}
 	}
 
+	// Validate and cache user-defined name transformations.
+	nts, err := reposource.CompileGitLabNameTransformations(c.NameTransformations)
+	if err != nil {
+		return nil, err
+	}
+
 	return &GitLabSource{
-		svc:     svc,
-		config:  c,
-		exclude: exclude,
-		baseURL: baseURL,
-		client:  gitlab.NewClientProvider(baseURL, cli).GetPATClient(c.Token, ""),
+		svc:                 svc,
+		config:              c,
+		exclude:             exclude,
+		baseURL:             baseURL,
+		nameTransformations: nts,
+		client:              gitlab.NewClientProvider(baseURL, cli).GetPATClient(c.Token, ""),
 	}, nil
 }
 
@@ -100,11 +108,13 @@ func (s GitLabSource) makeRepo(proj *gitlab.Project) *Repo {
 			s.config.RepositoryPathPattern,
 			s.baseURL.Hostname(),
 			proj.PathWithNamespace,
+			s.nameTransformations,
 		)),
 		URI: string(reposource.GitLabRepoName(
 			"",
 			s.baseURL.Hostname(),
 			proj.PathWithNamespace,
+			s.nameTransformations,
 		)),
 		ExternalRepo: gitlab.ExternalRepoSpec(proj, *s.baseURL),
 		Description:  proj.Description,