+++ pkg/conf/reposource/common_test.go
@@ -153,3 +153,40 @@ func TestParseCloneURL(t *testing.T) {
 		}
 	}
 }
+
+func TestNameTransformations(t *testing.T) {
+	opts := []NameTransformationOptions{
+		{
+			Regex:       `\.d/`,
+			Replacement: "/",
+		},
+		{
+			Regex:       "-git$",
+			Replacement: "",
+		},
+	}
+
+	nts := make([]NameTransformation, len(opts))
+	for i, opt := range opts {
+		nt, err := NewNameTransformation(opt)
+		if err != nil {
+			t.Fatalf("NewNameTransformation: %v", err)
+		}
+		nts[i] = nt
+	}
+
+	tests := []struct {
+		input  string
+		output string
+	}{
+		{"path/to.d/repo-git", "path/to/repo"},
+		{"path/to.d/repo-git.git", "path/to/repo-git.git"},
+		{"path/to.de/repo-git.git", "path/to.de/repo-git.git"},
+	}
+	for _, test := range tests {
+		got := NameTransformations(nts).Transform(test.input)
+		if test.output != got {
+			t.Errorf("for input %s, expected %s, but got %s", test.input, test.output, got)
+		}
+	}
+}